revised staircase method

+++ Source/ColorCalibration/Private/CalibratedCamera.cpp
@@ -37,7 +37,7 @@ void ACalibratedCamera::BeginPlay()
 	Normal_settings = alterCamera->PostProcessSettings;
 }
 
-void ACalibratedCamera::initialize_color_assist(FString primaries_filename, FString plates_filename, float start_threshold, float start_step_size, TArray<FTransform>& all_plates_transform) {
+void ACalibratedCamera::initialize_color_assist(FString primaries_filename, FString plates_filename, float start_threshold, int start_step_size, TArray<FTransform>& all_plates_transform) {
 	TArray<FColor_lxy> primaries;
 	cal_lib->readPrimariesFromCSV(primaries_filename, primaries);
 	cal_lib->readPlatePointsFromCSV(plates_filename, start_threshold, start_step_size, all_plates_transform);
